Do not send mails every 30 minutes

+++ src/WebsiteApi/NotificationsBundle/Command/NotificationMailCommand.php
@@ -9,6 +9,7 @@ use Symfony\Component\Process\Process;
 use WebsiteApi\NotificationsBundle\Entity\MailNotificationQueue;
 use Emojione\Client;
 use Emojione\Ruleset;
+use WebsiteApi\NotificationsBundle\Entity\UserNotificationStatus;
 use WebsiteApi\UsersBundle\Entity\User;
 
 
@@ -38,8 +39,21 @@ class NotificationMailCommand extends ContainerAwareCommand
 
             if ($date > 60 * 30) //30 minutes
             {
-                $users[] = Array("user" => $entry->getUserId(), "date" => $entry->getDate()->getTimestamp());
-                $em->remove($entry);
+
+                $user_notification_status = $em->getRepository("TwakeNotificationsBundle:UserNotificationStatus")->findOneBy(Array("user_id" => $user_id));
+                if (!$user_notification_status) {
+                    $user_notification_status = new UserNotificationStatus($user);
+                }
+
+                if ($user_notification_status->getMailStatus() == 0) {
+                    $user_notification_status->setMailStatus(1);
+
+                    $users[] = Array("user" => $entry->getUserId(), "date" => $entry->getDate()->getTimestamp());
+                    $em->remove($entry);
+
+                    $em->persist($user_notification_status);
+                }
+
             }
 
             error_log($date);
